Created RecipeDetails page and other supporting components.
Created RecipeDetails page which shows the ingredient list, instructions and a picture of the item. Will need to add additional components in the next update

Also added a different scroll bar that is a bit more modern looking

Removed footer from the HomePage file and instead included it in the App file so that it loads on every page.