Refresh token does not work
This would typically keep our user constantly relogging in; however it is not working; we are not receiving any secretCookies.

As this functionality is not necessary for an MVP, work will be postponed for now. We will still create them, but they are automatically removed at logout. The API call will be disabled to speed up the app