Admin can now delete a recipe
Admin users can now click to delete a recipe, in which they are prompted with a modal to confirm. Once confirmed, the recipe is deleted and page refreshed.

Fixed an issue when clicking logout if you dont have a valid session with the server; if we return undefined from the server (ie, we fail to  authenticate because the session ended), just end what the app considers a session as well