Refactored a bit to clean code up. Updated recipe view
Instead of being in a row, all view besides admin view will now look like a card. Instead of showing the recipeBook image, we now just mention that it is part of that recipe book as a text.

Removed the view button; Clicking on the recipe now brings you to the recipe page. When hovering over it, the card will expand and a box shadow will appear. When clicking on it, the background will darken. These will help indicate that something can be done to the user.