RecipeAddPage work finished
RecipeAddPage (which is also used for editing) has been completed. Some additional work was done so that the code behaves accordingly:

App.Js updated to use the new page. Old one has been renamed as such.

RecipeAddPage completed; Users can now click on the sections they want to edit in order to edit that text. This will need to be enhanced so that the user understands that they are still editing the page.

RecipeDetailsPage updated to now show the title of the book the recipe is in, desc of the recipe as well as the tags; these were not shown previously.

RecipePage now shows an add button if you are an admin user

IngredientList has been updated to no longer iterate over an array; we now pass in a string and break that up by the enter keys found in the string

FormattedSteps no longer shows a step number

RecipeDetails has the same updates as IngredientList

Removed unnecessary items from NavCanvas.js (an unused file_