UI changes to formatting, updates to app with searching
Small adjustments to UI with formatting.

Updated the bottomNav with a drop down that allows you to immediately filter results by certain options

To allow the above, the app now manages what is being searched via searchContext, and based off those values, the app will be able to filter results.