Renamed Alert file to AlertDisplay, fixed refresh bug
Renamed the Alert file to AlertDisplay and updated appropriate files to the new path.

Fixed bug where AlertDisplay wouldn't return anything if you were passing ing the lg prop

Fixed an issue with the RecipeReview where if you deleted a review, it prompted you to immediately add a new one, which isn't necessarily the case.

Fixed a bug where submitting or deleting a review always brought you to the top of the page; now it will just bring you to the top when the id changes