NavBar search UI has been implemented
First, this now lives in its own component.

Clicking to search from the options now hides the other options and just shows an input to allow you to search, or cancel and hide the and show the other options. From there, the user can pick either products or recipes, but not both