Update to user page, navbar and links throughout website
Both BookDiv and BookCard now direct to the books in the shop rather than going nowhere/ being disabled

LoginForm css updated so it behaves nicely on smaller screens

UserPage layout started. Welcomes the user and they are able to toggle between likes and settings

BottomNav updated with a test using the accordion style drop down instead. currently staying with the old layout

NavBar updated so it shows the login / logout, profile or admin options and the cart values